retornando avaliação certa e backup do bd

+++ database/migrations/2025_05_26_112039_create_avaliacao_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('comentario');
-            $table->unsignedTinyInteger('estrelas');
+            $table->double('estrelas');
             $table->unsignedBigInteger('alvo_id');
             $table->unique(['user_id', 'alvo_id']);
             $table->timestamps();
